Remove unnecessary AuthenticationInput component
The AuthenticationInput component was removed completely as it was no longer being used due to refactor and streamlining of the authentication system. The AppState type was also added to manage the global state of the application. In the App.tsx, the new AppState type was integrated and more context for the state of the app is now being provided for cleaner data management.

Furthermore, localization support is introduced by creating a translations.ts file to store all translations strings. This change will make the app more accessible to users from different locales. The translations are integrated in the AuthenticationPage as well.

Overall, these changes streamline the authentication process, improve the structure of the app's global state, and add support for localization.